closes #278. support more functonality in internal api.
