+++ demo/config/targets/INST/screens/tabs.txt
@@ -38,7 +38,6 @@ TABBOOK
 
     VERTICALBOX
       LABELVALUE INST ADCS TIMEFORMATTED WITH_UNITS 30
-      LABELVALUE COSMOS VERSION COSMOS
     END
   END
 
@@ -62,7 +61,6 @@ TABBOOK
 
     VERTICALBOX
       LABELVALUE INST HEALTH_STATUS TIMEFORMATTED WITH_UNITS 30
-      LABELVALUE COSMOS VERSION COSMOS
     END
   END
 END
