+++ demo/config/tools/test_runner/test_runner.txt
@@ -1,4 +1,4 @@
-LOAD_UTILITY 'example_test'
+REQUIRE_UTILITY 'example_test'
 LINE_DELAY 0
 # RESULTS_WRITER 'results_writer.rb'
 ALLOW_DEBUG
