+++ lib/cosmos/script/telemetry.rb
@@ -195,25 +195,6 @@ module Cosmos
       results
     end
 
-    # Get a packet which was previously subscribed to by
-    # subscribe_packet_data. This method can block waiting for new packets or
-    # not based on the second parameter. It returns a single Cosmos::Packet instance
-    # and will return nil when no more packets are buffered (assuming non_block
-    # is false).
-    # Usage:
-    #   get_packet(id, <true or false to block>)
-    def get_packet(id, non_block = false)
-      packet = nil
-      buffer, target_name, packet_name, received_time, received_count = get_packet_data(id, non_block)
-      if buffer
-        packet = System.telemetry.packet(target_name, packet_name).clone
-        packet.buffer = buffer
-        packet.received_time = received_time
-        packet.received_count = received_count
-      end
-      packet
-    end
-
   end
 end
