+++ lib/cosmos/script/tools.rb
@@ -14,70 +14,6 @@ module Cosmos
     private
 
     #######################################
-    # Methods accessing tlm_viewer
-    #######################################
-
-    def display(display_name, x_pos = nil, y_pos = nil)
-      run_tlm_viewer("display", display_name) do |tlm_viewer|
-        tlm_viewer.display(display_name, x_pos, y_pos)
-      end
-    end
-
-    def clear(display_name)
-      run_tlm_viewer("clear", display_name) do |tlm_viewer|
-        tlm_viewer.clear(display_name)
-      end
-    end
-
-    def clear_all(target = nil)
-      run_tlm_viewer("clear_all") do |tlm_viewer|
-        tlm_viewer.clear_all(target)
-      end
-    end
-
-    def run_tlm_viewer(action, display_name = '')
-      tlm_viewer = JsonDRbObject.new System.connect_hosts['TLMVIEWER_API'], System.ports['TLMVIEWER_API']
-      begin
-        yield tlm_viewer
-        tlm_viewer.disconnect
-      rescue DRb::DRbConnError
-        # No Listening Tlm Viewer - So Start One
-        start_tlm_viewer
-        max_retries = 60
-        retry_count = 0
-        begin
-          yield tlm_viewer
-          tlm_viewer.disconnect
-        rescue DRb::DRbConnError
-          retry_count += 1
-          if retry_count < max_retries
-            canceled = cosmos_script_sleep(1)
-            retry unless canceled
-          end
-          raise "Unable to Successfully Start Listening Telemetry Viewer: Could not #{action} #{display_name}"
-        rescue Errno::ENOENT
-          raise "Display Screen File: #{display_name}.txt does not exist"
-        end
-      rescue Errno::ENOENT
-        raise "Display Screen File: #{display_name}.txt does not exist"
-      end
-    end
-
-    def start_tlm_viewer
-      system_file = File.basename(System.initial_filename)
-      mac_app = File.join(Cosmos::USERPATH, 'tools', 'mac', 'TlmViewer.app')
-
-      if Kernel.is_mac? && File.exist?(mac_app)
-        Cosmos.run_process("open '#{mac_app}' --args --system #{system_file}")
-      else
-        cmd = 'ruby'
-        cmd << 'w' if Kernel.is_windows? # Windows uses rubyw to avoid creating a DOS shell
-        Cosmos.run_process("#{cmd} '#{File.join(Cosmos::USERPATH, 'tools', 'TlmViewer')}' --system #{system_file}")
-      end
-      cosmos_script_sleep(1)
-    end
-
-    #######################################
     # Methods accessing script runner
     #######################################
